⚡️ feat: 表示に関するFormatクラス実装
