+++ 07.ls_object/ls.rb
@@ -1,7 +1,8 @@
 # frozen_string_literal: true
 
 require_relative 'filelist'
+require_relative 'format'
 
 option = Option.new.extract_entered_option
 fl = FileList.new(option)
-fl.show(option)
+Format.new(option, fl.files).show
